Move the storage details to a separate module
This PR moves the rkv implementation details,
including Lifetime.Application handling, to
the newly introduced Database module.
